requirements.txt
libtiff==0.4.2
pandas==0.25.1
svgutils==0.3.1
scipy==1.3.1
tqdm==4.32.1
h5py==2.8.0
matplotlib==3.1.1
cmocean==2.0
scikit_image==0.15.0
lxml==4.4.1
keras_contrib==2.0.8
Keras==2.2.4
numpy==1.16.5
imageio==2.8.0
Pillow==7.1.1
pygame==1.9.6
PyOpenGL==3.1.5
skimage==0.0
